+++ core/sql/executor/ExExeUtilExplain.cpp
@@ -237,8 +237,7 @@ short ExExeUtilDisplayExplainTcb::work()
 	      executeImmediate("control session 'EXPLAIN' 'ON';");
 	    if (retcode < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		pstate.step_ = HANDLE_ERROR_;
 		break;
 	      }
@@ -250,8 +249,7 @@ short ExExeUtilDisplayExplainTcb::work()
 					 "__EXPL_STMT_NAME__");
 	    if (retcode < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		pstate.step_ = HANDLE_ERROR_;
 		break;
 	      }
@@ -274,8 +272,7 @@ short ExExeUtilDisplayExplainTcb::work()
 	    
 	    if (retcode < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		pstate.step_ = HANDLE_ERROR_;
 		break;
 	      }
@@ -365,8 +362,7 @@ short ExExeUtilDisplayExplainTcb::work()
 	    retcode = cliInterface()->fetchRowsPrologue(explainQuery_);
 	    if (retcode < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		pstate.step_ = HANDLE_ERROR_;
 		break;
 	      }
@@ -393,8 +389,7 @@ short ExExeUtilDisplayExplainTcb::work()
 	    retcode = cliInterface()->fetch();
 	    if (retcode < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		pstate.step_ = HANDLE_ERROR_;
 		break;
 	      }
@@ -457,8 +452,7 @@ short ExExeUtilDisplayExplainTcb::work()
 	    retcode = cliInterface()->fetch();
 	    if (retcode < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		pstate.step_ = HANDLE_ERROR_;
 		break;
 	      }
@@ -573,7 +567,7 @@ short ExExeUtilDisplayExplainTcb::work()
 	    retcode = cliInterface()->fetchRowsEpilogue(explainQuery_);
 	    if (retcode < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 	      }
 	    
 	    retcode = 
@@ -581,7 +575,7 @@ short ExExeUtilDisplayExplainTcb::work()
 					   input_desc_, output_desc_);
 	    if (retcode < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 	      }
 
 	    if (exeUtilTdb().getStmtName() == NULL)
@@ -599,7 +593,7 @@ short ExExeUtilDisplayExplainTcb::work()
 	      executeImmediate("control session reset 'EXPLAIN';");
 	    if (retcode < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 	      }
 	    
 	    if (pstate.step_ == FETCH_EPILOGUE_AND_RETURN_ERROR_)
@@ -3922,7 +3916,7 @@ short ExExeUtilDisplayExplainComplexTcb::work()
 		// Explain will return 'info not available' in its
 		// description field.
 		// All other errors are reported.
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		if (exeUtilTdb().loadIfExists() &&
 		     getDiagsArea()->contains(-1055))
 		  {
@@ -3992,8 +3986,7 @@ short ExExeUtilDisplayExplainComplexTcb::work()
 					 FALSE, exeUtilTdb().isVolatile());
 	    if (cliRC < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-		
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		step_ = DROP_AND_ERROR_;
 		break;
 	      }
@@ -4217,8 +4210,7 @@ short ExExeUtilDisplayExplainShowddlTcb::work()
 
 		if (cliRC < 0)
 		  {
-		    cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-		    
+                    setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		    step_ = ERROR_;
 		    break;
 		  }
@@ -4234,8 +4226,7 @@ short ExExeUtilDisplayExplainShowddlTcb::work()
 
 		if (cliRC < 0)
 		  {
-		    cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-		    
+                    setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		    step_ = ERROR_;
 		    break;
 		  }
@@ -4260,8 +4251,7 @@ short ExExeUtilDisplayExplainShowddlTcb::work()
 
 	    if (cliRC < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		step_ = ERROR_;
 		break;
 	      }
