+++ core/sql/executor/ExExeUtilMisc.cpp
@@ -102,7 +102,7 @@ short ExExeUtilFastDeleteTcb::doPurgedataCat(char * stmt)
   cliRC = holdAndSetCQD("EXE_PARALLEL_PURGEDATA", "OFF");
   if (cliRC < 0)
     {
-      cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+      setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
       return -1;
     }
 
@@ -111,8 +111,7 @@ short ExExeUtilFastDeleteTcb::doPurgedataCat(char * stmt)
 
   if (cliRC < 0)
     {
-      cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-
+      setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
       restoreCQD("EXE_PARALLEL_PURGEDATA");
       return -1;
     }
@@ -157,7 +156,7 @@ short ExExeUtilFastDeleteTcb::doLabelPurgedata(char * objectName,
 
   if (cliRC < 0)
     {
-      cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+      setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
       retcode = -1;
       goto cleanUpAndReturn;
     }
@@ -215,7 +214,7 @@ short ExExeUtilFastDeleteTcb::doFastDelete(char * objectName,
     executeImmediate("control query shape hold;");
   if (cliRC < 0)
     {
-      cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+      setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
       retcode = -1;
       goto cleanUpAndReturn;
     }
@@ -224,7 +223,7 @@ short ExExeUtilFastDeleteTcb::doFastDelete(char * objectName,
     cliInterface()->executeImmediate("control query shape esp_exchange(cut);");
   if (cliRC < 0)
     {
-      cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+      setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
       retcode = -1;
       goto cleanUpAndReturn;
     }
@@ -236,7 +235,7 @@ short ExExeUtilFastDeleteTcb::doFastDelete(char * objectName,
 	executeImmediate("control query default mv_internal_ignore_uninitialized 'ON';");
       if (cliRC < 0)
 	{
-	  cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+          setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 	  retcode = -1;
 	  goto cleanUpAndReturn;
 	}
@@ -276,7 +275,7 @@ short ExExeUtilFastDeleteTcb::doFastDelete(char * objectName,
 
   if (cliRC < 0)
     {
-      cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+      setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
       retcode = -1;
       goto cleanUpAndReturn;
     }
@@ -333,14 +332,14 @@ short ExExeUtilFastDeleteTcb::injectError(const char * val)
   if (((e1) && (strcmp(e1, val) == 0)) ||
       ((e2) && (strcmp(e2, val) == 0)))
     {
-      ComDiagsArea * da = getDiagsArea();
-      *da << DgSqlCode(-EXE_ERROR_INJECTED)
-	  << (e1 ? DgString0("SQLMX_TEST_POINT") : DgString0("SQLMX_PPD_ERR_TEST_POINT"))
-	  << DgInt0((Lng32)str_atoi(val, strlen(val)));
-
+      Lng32 errNumParam = ((Lng32)str_atoi(val, strlen(val)));
+      ComDiagsArea * diagsArea = getDiagsArea();
+      ExRaiseSqlError(getHeap(), &diagsArea, -EXE_ERROR_INJECTED,
+          &errNumParam, NULL, NULL,   
+	  (e1 ? "SQLMX_TEST_POINT" : "SQLMX_PPD_ERR_TEST_POINT"));
+      setDiagsArea(diagsArea);
       return -EXE_ERROR_INJECTED;
     }
-  
   return 0;
 }
 
@@ -409,7 +408,7 @@ short ExExeUtilFastDeleteTcb::work()
 		    cliRC = cliInterface()->beginWork();
 		    if (cliRC < 0)
 		      {
-			cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                        setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 			step_ = ERROR_;
 			break;
 		      }
@@ -444,8 +443,7 @@ short ExExeUtilFastDeleteTcb::work()
 	      {
 		if (cliRC < 0)
 		  {
-		    cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-		    
+                    setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		    step_ = ERROR_;
 		    break;
 		  }
@@ -458,7 +456,7 @@ short ExExeUtilFastDeleteTcb::work()
 	    cliRC = cliInterface()->beginWork();
 	    if (cliRC < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		step_ = ERROR_;
 		break;
 	      }
@@ -508,8 +506,7 @@ short ExExeUtilFastDeleteTcb::work()
 	      rc = injectError("10");
 	    if (rc < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		// could not acquire a ddl lock.
 		// try to purgedata using catman sequential purgedata.
 
@@ -552,8 +549,7 @@ short ExExeUtilFastDeleteTcb::work()
 	      }
 	    if (rc < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		// security validation failed
 		if ((getDiagsArea()->contains(-1017)) ||
 		    (rollbackNoPDErr))
@@ -576,7 +572,7 @@ short ExExeUtilFastDeleteTcb::work()
 	      rc = injectError("12");
 	    if (rc < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		step_ = ROLLBACK_WORK_AND_ERROR_;
 		break;
 	      }
@@ -592,7 +588,7 @@ short ExExeUtilFastDeleteTcb::work()
 		  {
 		    strcpy(failReason_, "Error during commit work.");
 
-		    cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                    setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		    step_ = ROLLBACK_WORK_AND_ERROR_;
 		    break;
 		  }
@@ -607,7 +603,7 @@ short ExExeUtilFastDeleteTcb::work()
 		strcpy(buf, "SELECT TESTEXIT;");
 		cliRC = cliInterface()->executeImmediate(buf);
 		if (cliRC < 0)
-		  cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                  setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		  
 		strcpy(failReason_, " ");
 
@@ -625,7 +621,7 @@ short ExExeUtilFastDeleteTcb::work()
 	    cliRC = cliInterface()->beginWork();
 	    if (cliRC < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		step_ = ERROR_;
 		break;
 	      }
@@ -733,14 +729,13 @@ short ExExeUtilFastDeleteTcb::work()
 
 	    if (step_ == ROLLBACK_WORK_AND_ERROR_)
 	      {
-		ComDiagsArea * da = getDiagsArea();
-		
+		ComDiagsArea * diagsArea = getDiagsArea();
 		// convert all errors into warnings
-		NegateAllErrors(da);
-		
-		// add a new error to indicate that parallel purgedata failed.
-		*da << DgSqlCode(-EXE_PARALLEL_PURGEDATA_FAILED)
-		    << DgString0(failReason_);
+		NegateAllErrors(diagsArea);
+                ExRaiseSqlError(getHeap(), &diagsArea, -EXE_PARALLEL_PURGEDATA_FAILED,
+                    NULL, NULL, NULL,
+		    failReason_);
+                setDiagsArea(diagsArea);
 	      }
 
 	    step_ = ERROR_;
@@ -763,7 +758,7 @@ short ExExeUtilFastDeleteTcb::work()
 		strcpy(buf, "SELECT TESTEXIT;");
 		cliInterface()->executeImmediate(buf);
 		if (cliRC < 0)
-		  cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                  setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		
 		step_ = ERROR_;
 		break;
@@ -774,10 +769,11 @@ short ExExeUtilFastDeleteTcb::work()
 		// raise a warning that parallel purgedata was performed.
 		if (fdTdb().returnPurgedataWarn())
 		  {
-		    ComDiagsArea * da = getDiagsArea();
-		    *da << DgSqlCode(EXE_PURGEDATA_CAT)
-			<< DgString0("Parallel")
-			<< DgString1("");
+		    ComDiagsArea * diagsArea = getDiagsArea();
+                    ExRaiseSqlError(getHeap(), &diagsArea, EXE_PURGEDATA_CAT,
+                        NULL, NULL, NULL,
+			"Parallel", "");
+                    setDiagsArea(diagsArea);
 		  }
 		step_ = DONE_;
 	      }
@@ -791,7 +787,7 @@ short ExExeUtilFastDeleteTcb::work()
 	    cliRC = cliInterface()->beginWork();
 	    if (cliRC < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		step_ = ERROR_;
 		break;
 	      }
@@ -806,7 +802,7 @@ short ExExeUtilFastDeleteTcb::work()
 
 	    if (rc < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		step_ = KILL_MXCMP_AND_ERROR_;
 		break;
 	      }
@@ -823,7 +819,7 @@ short ExExeUtilFastDeleteTcb::work()
 	      rc = injectError("16");
 	    if (rc < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		step_ = KILL_MXCMP_AND_ERROR_;
 		break;
 	      }
@@ -842,7 +838,7 @@ short ExExeUtilFastDeleteTcb::work()
 		  rc = injectError("17");
 		if (rc < 0)
 		  {
-		    cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                    setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		    step_ = KILL_MXCMP_AND_ERROR_;
 		    break;
 		  }
@@ -851,7 +847,7 @@ short ExExeUtilFastDeleteTcb::work()
 	    cliRC = cliInterface()->commitWork();
 	    if (cliRC < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		step_ = ERROR_;
 		break;
 	      }
@@ -861,10 +857,11 @@ short ExExeUtilFastDeleteTcb::work()
 	    // raise a warning that parallel purgedata was performed.
 	    if (fdTdb().returnPurgedataWarn())
 	      {
-		ComDiagsArea * da = getDiagsArea();
-		*da << DgSqlCode(EXE_PURGEDATA_CAT)
-		    << DgString0("Parallel")
-		    << DgString1("");
+		ComDiagsArea * diagsArea = getDiagsArea();
+                ExRaiseSqlError(getHeap(), &diagsArea, EXE_PURGEDATA_CAT,
+                    NULL, NULL, NULL,
+		    "Parallel", "");
+                setDiagsArea(diagsArea);
 	      }
 	  }
 	break;
@@ -883,15 +880,16 @@ short ExExeUtilFastDeleteTcb::work()
 	    // for parallel purgedata, or because parallel purgedata failed.
 	    if (fdTdb().returnPurgedataWarn())
 	      {
-		ComDiagsArea * da = getDiagsArea();
+		ComDiagsArea * diagsArea = getDiagsArea();
 		if (fdTdb().doPurgedataCat())
-		  *da << DgSqlCode(EXE_PURGEDATA_CAT)
-		      << DgString0("Regular")
-		      << DgString1("Reason: Query or the object did not meet the criteria for parallel purgedata.");
+                  ExRaiseSqlError(getHeap(), &diagsArea, EXE_PURGEDATA_CAT,
+                     NULL, NULL, NULL,
+                     "Regular", "Reason: Query or the object did not meet the criteria for parallel purgedata.");
 		else
-		  *da << DgSqlCode(EXE_PURGEDATA_CAT)
-		      << DgString0("Regular")
-		      << DgString1("Reason: Parallel purgedata failed.");
+                  ExRaiseSqlError(getHeap(), &diagsArea, EXE_PURGEDATA_CAT,
+                     NULL, NULL, NULL,
+                     "Regular", "Reason: Parallel purgedata failed.");
+                setDiagsArea(diagsArea);
 	      }
 
 	    rc = doPurgedataCat(fdTdb().purgedataStmt());
@@ -1365,7 +1363,7 @@ short ExExeUtilLongRunningTcb::executeLongRunningQuery()
         "control query default HIST_ON_DEMAND_STATS_SIZE '0'");
       if (cliRC < 0) 
         {
-          cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+          setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 	  return cliRC;
         }
       
@@ -1374,7 +1372,7 @@ short ExExeUtilLongRunningTcb::executeLongRunningQuery()
       short rtc = processInitial(rc);
       if ((rc != 0) && (rc != 100))
         {
-          cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+          setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
         }
 
       cliRC = 
@@ -1382,7 +1380,7 @@ short ExExeUtilLongRunningTcb::executeLongRunningQuery()
             "control query default HIST_ON_DEMAND_STATS_SIZE 'RESET'");
       if (cliRC < 0) 
         {
-          cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+          setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 	  return cliRC;
         }
       
@@ -1418,7 +1416,7 @@ short ExExeUtilLongRunningTcb::executeLongRunningQuery()
       short rtc = processContinuing(rc);
       if ((rc != 0) && (rc != 100))
         {
-          cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
+          setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
         }
 
 #ifdef _DEBUG
@@ -1980,8 +1978,7 @@ short ExExeUtilPopulateInMemStatsTcb::work()
 	    
 	    if (cliRC < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		step_ = ERROR_;
 	      }
 	    else
@@ -2037,8 +2034,7 @@ short ExExeUtilPopulateInMemStatsTcb::work()
 	    
 	    if (cliRC < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		step_ = ERROR_;
 	      }
 	    else
@@ -2094,8 +2090,7 @@ short ExExeUtilPopulateInMemStatsTcb::work()
 	    
 	    if (cliRC < 0)
 	      {
-		cliInterface()->retrieveSQLDiagnostics(getDiagsArea());
-
+                setDiagsArea(cliInterface()->allocAndRetrieveSQLDiagnostics(getDiagsArea()));
 		step_ = ERROR_;
 	      }
 	    else
@@ -2256,18 +2251,13 @@ ExExeUtilHiveTruncateTcb::~ExExeUtilHiveTruncateTcb()
 void ExExeUtilHiveTruncateTcb::freeResources()
 {
   if (htTdb().getDropOnDealloc())
-    {
+  {
       NAString hiveDropDDL("drop table ");
-      HiveClient_JNI *hiveClient = HiveClient_JNI::getInstance();
-
       hiveDropDDL += htTdb().getHiveTableName();
 
-      // ignore errors on drop
-      if (!hiveClient->isInitialized() ||
-          !hiveClient->isConnected())
-        hiveClient->init();
-      hiveClient->executeHiveSQL(hiveDropDDL);
-    }
+      // TODO: is it ok to ignore the error 
+      HiveClient_JNI::executeHiveSQL(hiveDropDDL);
+  }
   if (lobGlob_) {
     ExpLOBinterfaceCleanup(lobGlob_);
     lobGlob_ = NULL;
@@ -2426,7 +2416,6 @@ short ExExeUtilHiveTruncateTcb::work()
                           NULL, NULL, NULL,
                           reason,
                           NULL, NULL);
-          
           pentry_down->setDiagsArea(diagsArea);
           step_ = ERROR_;
         }
@@ -2572,15 +2561,16 @@ short ExExeUtilHiveQueryTcb::work()
           
         case PROCESS_QUERY_:
           {
-            ComDiagsArea * diags = getDiagsArea();
-
-            if (!CmpCommon::context()->execHiveSQL(htTdb().getHiveQuery(),
-                                                   diags))
-              {
+            if (HiveClient_JNI::executeHiveSQL(htTdb().getHiveQuery()) != HVC_OK)
+            {
+                ComDiagsArea * diagsArea = NULL;
+                ExRaiseSqlError(getHeap(), &diagsArea, -1214,
+                        NULL, NULL, NULL,
+                        getSqlJniErrorStr(), htTdb().getHiveQuery()); 
+                setDiagsArea(diagsArea);
                 step_ = ERROR_;
                 break;
-              }
-
+            }
             step_ = DONE_;
           }
           break;
@@ -2589,8 +2579,8 @@ short ExExeUtilHiveQueryTcb::work()
           {
             if (handleError())
               return WORK_OK;
-
-	    getDiagsArea()->clear();
+            if (getDiagsArea() != NULL)
+	       getDiagsArea()->clear();
             
             step_ = DONE_;
           }
